backend/requirements.txt
Flask
Flask_Cors
matplotlib
numpy
opencv_contrib_python
pandas
Pillow
ultralytics
